+++ src/_build/util.py
@@ -11,21 +11,19 @@ EXT_PLATFORM_SOURCE_CODE={"posix":"src/sll_ext/platform/posix","nt":"src/sll_ext
 
 
 
-log=(print if "--verbose" in sys.argv else lambda *_,**__:None)
+log=(print if "--verbose" in sys.argv else lambda *_:None)
 
 
 
 def fix_env():
 	if (os.name=="nt"):
-		if ("__ghactions" in sys.argv):
+		if (os.getenv("GITHUB_ACTIONS",None) is not None):
 			for k in str(subprocess.run([str(subprocess.run([os.environ["ProgramFiles(x86)"]+"/Microsoft Visual Studio/Installer/vswhere.exe","-nologo","-latest","-products","*","-requires","Microsoft.VisualStudio.Component.VC.Tools.x86.x64","-property","installationPath"],stdout=subprocess.PIPE).stdout.strip(),"utf-8")+"/VC/Auxiliary/Build/vcvarsall.bat","x64","&&","cls","&&","set"],shell=True,stdout=subprocess.PIPE).stdout.split(b"\x0c")[1],"utf-8").split("\r\n"):
 				k=[e.strip() for e in k.split("=")]
 				if (k[0].lower() in ["path","include","lib","libpath"]):
 					os.environ[k[0].upper()]=k[1]
 	else:
-		ld=os.getenv("LD_LIBRARY_PATH","")
-		if ("build:" not in ld):
-			os.environ["LD_LIBRARY_PATH"]="build:"+ld
+		os.environ["LD_LIBRARY_PATH"]="build:"+os.getenv("LD_LIBRARY_PATH","")