+++ src/sll_ext/main.c
@@ -4,6 +4,13 @@
 
 
 
+static sll_object_t* _external_func(const sll_object_t*const* al,sll_arg_count_t all){
+	sll_file_write_string(sll_stdout,SLL_CHAR("External function call!\n"));
+	return SLL_FROM_INT(1);
+}
+
+
+
 __SLL_EXT_EXTERNAL sll_bool_t __sll_load(sll_version_t v,sll_bool_t f){
 	if (v!=SLL_EXT_VERSION){
 		sll_file_write_string(sll_stdout,SLL_CHAR("<ext-lib> Invalid version!\n"));
@@ -13,6 +20,7 @@ __SLL_EXT_EXTERNAL sll_bool_t __sll_load(sll_version_t v,sll_bool_t f){
 		sll_file_write_string(sll_stdout,SLL_CHAR("<ext-lib> Loaded extension library for v"SLL_EXT_VERSION_STRING"! (compilation)\n"));
 		return 1;
 	}
+	sll_register_internal_function(sll_current_runtime_data->ift,SLL_CHAR("sll_ext:test_func"),_external_func,0);
 	sll_file_write_string(sll_stdout,SLL_CHAR("<ext-lib> Loaded extension library for v"SLL_EXT_VERSION_STRING"!\n"));
 	return 1;
 }